style: restructure
